requirements.txt
# requirements.txt
playwright
python-dotenv
ollama
beautifulsoup4
asyncio